feat: Create AlbumFileCard component

+++ src/pages/album/SavePage.vue
@@ -1,13 +1,14 @@
 <script setup lang="ts">
+import AlbumFileCard from 'src/components/AlbumFileCard.vue';
 import FileDataUpload from 'src/components/FileDataUpload.vue';
 import PageTitle from 'src/components/PageTitle.vue';
+import useDialog from 'src/composables/useDialog';
 import useNotify from 'src/composables/useNotify';
 import useAlbumFileService from 'src/services/abumFile.service';
 import useAlbumService from 'src/services/album.service';
 import { AlbumFileType } from 'src/types/AbumFile.type';
 import { AlbumType } from 'src/types/Album.type';
-import { AlbumFileFileTypeEnum } from 'src/types/enums/AlbumFileFileType.enum';
-import { onMounted, ref, Ref } from 'vue';
+import { onMounted, provide, ref, Ref } from 'vue';
 import { useRoute } from 'vue-router';
 
 defineOptions({
@@ -26,6 +27,7 @@ const notify = useNotify();
 const album: Ref<AlbumType | null> = ref(null);
 const albumFiles: Ref<AlbumFileType[]> = ref([]);
 const theme: Ref<string> = ref('#eee');
+const dialogConfirmation = useDialog();
 
 const getAlbum = async () => {
   try {
@@ -48,9 +50,18 @@ const getAlbum = async () => {
 };
 
 const removeAlbumFile = async (id: number) => {
-  try {
+  dialogConfirmation.confirm().onOk(async () => {
     if (id) {
       await albumFileService.remove(id);
+      getAlbum();
+    }
+  });
+};
+
+const updateAlbumFile = async (id: number, title: string) => {
+  try {
+    if (id) {
+      await albumFileService.put(id, { title: title });
       getAlbum();
     }
   } catch (error: any) {
@@ -59,6 +70,9 @@ const removeAlbumFile = async (id: number) => {
     notify.error(message);
   }
 };
+
+provide('removeAlbumFile', removeAlbumFile);
+provide('updateAlbumFile', updateAlbumFile);
 </script>
 
 <template>
@@ -82,7 +96,7 @@ const removeAlbumFile = async (id: number) => {
         >
       </q-card-section>
       <q-card-section>
-        <FileDataUpload />
+        <FileDataUpload @uploaded="getAlbum" />
       </q-card-section>
       <q-card-section>
         <div class="column q-gutter-y-md">
@@ -92,61 +106,7 @@ const removeAlbumFile = async (id: number) => {
               v-for="(albumFile, index) in albumFiles"
               :key="index"
             >
-              <q-card
-                :style="{
-                  backgroundColor: theme,
-                  border: '10px solid' + theme,
-                }"
-                dark
-              >
-                <video
-                  class="full-width"
-                  v-if="albumFile.file_type === AlbumFileFileTypeEnum.VIDEO"
-                  :src="albumFile.file_path"
-                  :autoplay="false"
-                  style="height: 200px"
-                  controls
-                />
-                <q-img
-                  v-else
-                  :src="`${albumFile.file_path}`"
-                  style="height: 200px"
-                />
-                <q-btn
-                  flat
-                  round
-                  :style="{ backgroundColor: theme }"
-                  :icon="
-                    albumFile.file_type === AlbumFileFileTypeEnum.VIDEO
-                      ? 'las la-video'
-                      : 'las la-image'
-                  "
-                  class="absolute"
-                  style="top: 13px; right: -5px; transform: translateY(-50%)"
-                />
-                <q-card-section>
-                  <div class="text-h6">
-                    {{ albumFile.title ? albumFile.title : '&nbsp;' }}
-                  </div>
-                </q-card-section>
-                <q-card-actions>
-                  <div class="text-body2 text-right">
-                    <q-chip
-                      color="warning"
-                      text-color="black"
-                      icon="las la-cloud-upload-alt"
-                    >
-                      {{ albumFile.memory_usage }}
-                    </q-chip>
-                  </div>
-                  <q-btn
-                    @click="removeAlbumFile(albumFile.id)"
-                    icon="las la-trash"
-                    color="black"
-                    flat
-                  ></q-btn>
-                </q-card-actions>
-              </q-card>
+              <AlbumFileCard :album-file="albumFile" :theme="theme" />
             </div>
           </div>
         </div>
